pyproject.toml
[project]
name = "daily-snapshot-hub"
version = "0.1.0"
description = "GitHub Actions based daily data collection and archiving hub"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/daily_hub"]

[tool.pytest.ini_options]
testpaths = ["tests"]
